Vitest batch 24: test-custom-api (138 asserts split)
Split off:
  - test-custom-api.js (125 asserts) → Vitest. Source-inspection of api.js,
    settings.js, provider-panels.js, pdf-import.js, service-worker.js,
    api/proxy.js, crypto.js; behavioral tests for URL/key/model management,
    hasAIProvider gating (URL+key both required), callCustomAPI error paths
    (throws "No Custom API URL"/"No Custom API key" before any fetch),
    fetchCustomApiModels empty-arg early-returns, and the full
    needsMaxCompletionTokens GPT-5/o-series detection matrix (#114).

  - test-custom-api-dom.js (13 asserts) → new puppeteer file. Sections
    13/14: openSettingsModal provider grid, switchAIProvider rendering the
    Custom panel form fields, and the connected-state model dropdown
    (<select> with .options).

The section-19 runtime key-rehydrate probe was already gated on
window.isUnlocked() — in Node that's false, so only the crypto.js
source-string check (API_KEY_LS_KEYS includes labcharts-custom-key)
runs in Vitest; the puppeteer suite still exercises the runtime path.

Vitest: 66 → 67 tests, ~5.0s → ~4.45s.
Puppeteer: 34 → 33 remaining (test-custom-api swapped for slim DOM file).
Full run-tests.sh: green.
